feat: separation of business logic and display states
