+++ src/hooks/useDisplayState.ts
@@ -1,0 +1,21 @@
+import { useState, useEffect } from 'react';
+
+export const useDisplayState = () => {
+  const [currentPage, setCurrentPage] = useState(0);
+  const [windowWidth, setWindowWidth] = useState(window.innerWidth);
+
+  useEffect(() => {
+    const handleResize = () => {
+      setWindowWidth(window.innerWidth);
+    };
+    window.addEventListener('resize', handleResize);
+
+    return () => {
+      window.removeEventListener('resize', handleResize);
+    };
+  }, []);
+
+  return { currentPage, windowWidth, setCurrentPage };
+};
+
+export default useDisplayState;
